#153 post 테이블의 price,views 자료형 int로 변환 (#154)
* chore : toastUi 이미지 등록 툴바 미적용 수정

* feature : 댓글 수정 기능 추가

* fix : post의 list html파일 댓글 디자인 변경

* fix : 포스트 수정하기의 toast editor ui가 제대로 나타나지 않는 현상을 수정함

* chore: 시리즈 페이지 좋아요,조회수 배치 수정

* feature: 포스트 페이지에서는 각 포스트의 좋아요 갯수가 뜨도록 구현

* feature: 시리즈 페이지에서 포스트들의 좋아요 합으로 나타나도록 구현

* fix: 좋아요 DB의 postCode때문에 외래키 제약으로 삭제 되지 않던 현상을 수정

* feature: 포스트 페이지에서 상단 바에 좋아요 갯수 나타나도록 구현

* feature: 메인 페이지 인기 포스트 좋아요 보이도록 설정

* feature: 메인 페이지 인기 포스트 좋아요 순으로 5개 나타나도록 설정

* feature: 메인 페이지 인기 시리즈 좋아요 순으로 5개 나타나도록 설정

* feature: 메인 페이지 인기 시리즈 좋아요 순으로 10개 나타나도록 설정

* fix: post 테이블의 price,view 자료형 변환

+++ src/main/java/com/pentoryall/post/controller/PostController.java
@@ -90,7 +90,7 @@ public class PostController {
         char isPublic = params.get("isPublic") != null ? params.get("isPublic").charAt(0) : 'N';
         long seriesno = Long.parseLong(params.get("series"));
         char isFee = params.get("isFee") != null ? params.get("isFee").charAt(0) : 'N';
-        long neededPoint = Long.parseLong(params.get("neededPoint"));
+        int neededPoint = Integer.parseInt(params.get("neededPoint"));
         char isAdult = params.get("isAdult") != null ? params.get("isAdult").charAt(0) : 'N';
